+++ webpack/devices/connectivity/index.tsx
@@ -1,11 +1,15 @@
 import * as React from "react";
 import { Widget, WidgetHeader, WidgetBody, Row, Col } from "../../ui/index";
 import { t } from "i18next";
-import { CowardlyDictionary } from "../../util";
+import { ConnectivityRow, StatusRowProps } from "./connectivity_row";
+import { RetryBtn } from "./retry_btn";
+import { SpecialStatus } from "../../resources/tagged_resources";
 
 interface Props {
+  onRefresh(): void;
   rowData: StatusRowProps[];
   children?: React.ReactChild;
+  status: SpecialStatus | undefined;
 }
 
 interface State {
@@ -21,7 +25,10 @@ export class ConnectivityPanel extends React.Component<Props, State> {
       <WidgetHeader
         title={t("Connectivity")}
         helpText={t("Diagnose connectivity issues with FarmBot and the browser.")}>
-        <RetryBtn flags={rowData.map(x => !!x.connectionStatus)} />
+        <RetryBtn
+          status={this.props.status}
+          onClick={this.props.onRefresh}
+          flags={rowData.map(x => !!x.connectionStatus)} />
       </WidgetHeader>
       <WidgetBody>
         <ConnectivityRow from="from" to="to" />
@@ -36,53 +43,4 @@ export class ConnectivityPanel extends React.Component<Props, State> {
       </WidgetBody>
     </Widget>;
   }
-}
-
-/** Data model for a single row within the <ConnectivityPanel /> */
-export interface StatusRowProps {
-  connectionStatus?: boolean | undefined;
-  from: string;
-  to: string;
-  children?: React.ReactChild;
-}
-
-const iconLookup: CowardlyDictionary<string> = {
-  true: "saucer green active",
-  false: "saucer red active"
-};
-
-function ConnectivityRow(props: StatusRowProps) {
-  const className = iconLookup["" + props.connectionStatus] || "saucer yellow active";
-  return <Row>
-    <Col xs={1}>
-      <div className={className}></div>
-    </Col>
-    <Col xs={1}>
-      <p>
-        {props.from}
-      </p>
-    </Col>
-    <Col xs={1}>
-      <p>
-        {props.to}
-      </p>
-    </Col>
-    <Col xs={9}>
-      <p>
-        {props.children}
-      </p>
-    </Col>
-  </Row>;
-}
-
-interface RetryBtnProps {
-  flags: boolean[];
-}
-export function RetryBtn(props: RetryBtnProps) {
-  const failures = props.flags.includes(false);
-  const color = failures ? "red" : "green";
-
-  return <button className={color + " fb-button"}>
-    Check Again
-</button>;
 }
